fix: avoid exposing hook and change prop
