+++ packages/preset-dumi/src/theme/index.ts
@@ -14,7 +14,6 @@ export { default as useDemoUrl } from './hooks/useDemoUrl';
 export { default as useApiData } from './hooks/useApiData';
 export { default as useTSPlaygroundUrl } from './hooks/useTSPlaygroundUrl';
 export { default as usePrefersColor } from './hooks/usePrefersColor';
-export { default as useLinkHighlight } from './hooks/useLinkHighlight';
 
 export interface IPreviewerComponentProps {
   title?: string;
